add name change capability

+++ web-app/app/api/save/route.js
@@ -7,24 +7,29 @@ import { NextRequest, NextResponse } from "next/server";
  * @param { NextRequest } req
  */
 export async function POST(req) {
-  const { content, id } = await req.json();
-
-  const uniqueNonStopwords = getUniqueNonStopwords(content);
-
+  const { title, content, id } = await req.json();
   const loadedConnection = await connection;
 
-  await loadedConnection.beginTransaction();
-  await loadedConnection.query("DELETE FROM NotesIndex WHERE id = ?", [id]);
-
-  for (const { word, index } of uniqueNonStopwords) {
-    await loadedConnection.query(
-      "INSERT IGNORE INTO NotesIndex VALUES (?, ?, ?)",
-      [id, word, index]
-    );
+  if (content) {
+    const uniqueNonStopwords = getUniqueNonStopwords(content);
+  
+    await loadedConnection.beginTransaction();
+    await loadedConnection.query("DELETE FROM NotesIndex WHERE id = ?", [id]);
+  
+    for (const { word, index } of uniqueNonStopwords) {
+      await loadedConnection.query(
+        "INSERT IGNORE INTO NotesIndex VALUES (?, ?, ?)",
+        [id, word, index]
+      );
+    }
+    await (
+      await connection
+    ).query("UPDATE Notes SET content = ? WHERE id = ?;", [content, id]);
+    await loadedConnection.commit();
+  } else if (title) {
+    await loadedConnection.beginTransaction();
+    await loadedConnection.query("UPDATE Notes SET title = ? WHERE id = ?;", [title, id]);
+    await loadedConnection.commit();
   }
-  await (
-    await connection
-  ).query("UPDATE Notes SET content = ? WHERE id = ?;", [content, id]);
-  await loadedConnection.commit();
   return NextResponse.json({ message: "success" });
 }
